WIP Refactor retry()
`retry()` would be useful in the parent AWR package so I/"we" could reuse it on other AWS services - but first, I don't think we really need or want recursion in there. Also it would be nice to count down from `retries` rather than up to `3` hardcoded. Haven't tested this yet.